Remove UI related test cases on nextgen branch also remove test_db_performance.py and related docs
